+++ src/components/AsideHeader/useAsideHeaderInnerContextValue.tsx
@@ -4,7 +4,7 @@ import {MenuItem} from '../types';
 
 import {AsideHeaderInnerContextType} from './AsideHeaderContext';
 import {AllPagesPanel, getAllPagesMenuItem} from './components/AllPagesPanel';
-import {AsideHeaderItem, AsideHeaderProps, InnerPanels} from './types';
+import {AsideHeaderItem, AsideHeaderProps, InnerPanels, PanelItemProps} from './types';
 
 const EMPTY_MENU_ITEMS: AsideHeaderItem[] = [];
 
@@ -38,8 +38,8 @@ export const useAsideHeaderInnerContextValue = (
         Boolean(onMenuItemsChanged) && (!menuItems || menuItems?.length > 0);
 
     useEffect(() => {
-        // If any user panel became visible we need to switch off all inner panels
-        if (panelItems?.some((x) => x.visible)) {
+        // If any user panel became open we need to switch off all inner panels
+        if (panelItems?.some((x) => x.open)) {
             setInnerVisiblePanel(undefined);
         }
     }, [panelItems]);
@@ -98,16 +98,17 @@ export const useAsideHeaderInnerContextValue = (
         [allPagesIsAvailable, menuItems, innerVisiblePanel, ALL_PAGES_MENU_ITEM, onAllPagesClick],
     );
 
-    const innerPanelItems = useMemo(() => {
+    const innerPanelItems = useMemo<PanelItemProps[] | undefined>(() => {
         if (!allPagesIsAvailable) {
             return panelItems;
         }
+
         return [
             ...(panelItems || []),
             {
                 id: InnerPanels.AllPages,
-                content: <AllPagesPanel />,
-                visible: innerVisiblePanel === InnerPanels.AllPages,
+                children: <AllPagesPanel />,
+                open: innerVisiblePanel === InnerPanels.AllPages,
             },
         ];
     }, [allPagesIsAvailable, panelItems, innerVisiblePanel]);